requirements.txt
absl-py==0.9.0
appnope==0.1.0
asgiref==3.2.3
astor==0.8.1
attrs==19.1.0
awscli==1.16.226
backcall==0.1.0
bleach==3.1.4
botocore==1.12.216
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.3.9
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
future==0.18.2
gast==0.2.2
google-auth==1.10.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.8
image==1.5.27
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.14.1
json5==0.8.5
jsonschema==3.0.2
jupyter-client==5.3.3
jupyter-core==4.5.0
jupyterlab==1.1.3
jupyterlab-server==1.0.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
progressbar2==3.51.4
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.11.2
ptyprocess==0.6.0
pyasn1==0.4.6
pyasn1-modules==0.2.8
Pygments==2.4.2
pyparsing==2.4.6
pyrsistent==0.15.4
python-dateutil==2.8.0
python-utils==2.4.0
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==3.4.2
s3transfer==0.2.1
scikit-learn==0.22.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.11.2
